dz01(xml): added auth activity
