+++ app/src/main/res/values/colors.xml
@@ -5,5 +5,6 @@
     <color name="orange">#FB5E06</color>
     <color name="grayText">#18181F</color>
     <color name="grayText2">#B2B2B3</color>
-    <color name="Background">#E7E7E7</color>
+    <color name="background">#E7E7E7</color>
+    <color name="errorRed">#F93E3E</color>
 </resources>
