+++ app/src/main/res/values/integers.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <!-- =============== AUTHORISATION DIMENSIONS =============== -->
+    <integer name="password_length">8</integer>
+
+</resources>
