+++ app/src/main/res/values/strings.xml
@@ -16,7 +16,6 @@
     <!--  sign-up page  -->
     <string name="signup_title">Let\'s get acquainted</string>
     <string name="signup_subtitle">Fill out the profile and go to the application!</string>
-    <string name="email_error_text">Incorrect E-Mail address</string>
     <string name="signup_google">Google</string>
     <string name="signup_or">or</string>
     <string name="signup_register">Register</string>
@@ -34,5 +33,11 @@
     <string name="login">Login</string>
     <string name="dont_have_account">Don’t have account?</string>
     <string name="sign_up">Sign up</string>
+
+    <!--   login registration error text   -->
+    <string name="email_error_text">Incorrect E-Mail address</string>
+    <string name="password_error_text">Your password must include a minimum of 8 characters.</string>
+    <string name="user_name_error">User name must contain at least 3 letters</string>
+    <string name="phone_length_error">Phone must be at least 10 digits long</string>
 
 </resources>
